+++ DirectX11/graphics/InputLayout.cpp
@@ -6,7 +6,7 @@
 
 InputLayout::InputLayout(IRenderingContext& graphics, const std::vector<D3D11_INPUT_ELEMENT_DESC>& elements, ID3DBlob* blob)
 {
-    GIO_THROW_IF_FAILED(graphics.GetDevice()->CreateInputLayout(elements.data(), elements.size(), blob->GetBufferPointer(),
+    GIO_THROW_IF_FAILED(graphics.GetDevice()->CreateInputLayout(elements.data(),static_cast<UINT>(elements.size()), blob->GetBufferPointer(),
                                             blob->GetBufferSize(), &inputLayout));
     
 }
